MovieService/requirements.txt
asyncpg==0.27.0
databases[postgresql]==0.5.5
fastapi==0.95.0
SQLAlchemy==1.4.41
uvicorn==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6